+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## TODO
 None
 
-## [WIP] 0.2.0 / 2017-June-??
+## 0.2.0 / 2017-June-04
 ### Changed
 - Use `[bract.core "0.2.0"]`